CIF-2982 - Move extensibility to CategoryRetriever
