+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/models/productlist/ProductList.java
@@ -25,7 +25,6 @@ import org.osgi.annotation.versioning.ConsumerType;
 import com.adobe.cq.commerce.core.components.models.experiencefragment.CommerceExperienceFragmentContainer;
 import com.adobe.cq.commerce.core.components.models.page.PageMetadata;
 import com.adobe.cq.commerce.core.components.models.productcollection.ProductCollection;
-import com.adobe.cq.commerce.core.components.models.retriever.AbstractCategoryRetriever;
 import com.adobe.cq.commerce.core.components.storefrontcontext.CategoryStorefrontContext;
 import com.adobe.cq.wcm.core.components.models.Component;
 
@@ -71,7 +70,7 @@ public interface ProductList extends Component, ProductCollection, PageMetadata 
      *
      * @return category retriever instance
      */
-    AbstractCategoryRetriever getCategoryRetriever();
+    CategoryRetriever getCategoryRetriever();
 
     /**
      * The version 1 of the productlist component always returns <code>false</code> as it does not support this feature.
